don't add indent level for include as a resource parameter (closes: #99)
The current code was designed to fit the case where one would write
this:

    include class1,
      class2,
      class3

But it creates a problem with resources that might have a parameter
named include. For example in the following situation, we would have an
indent added even though there's no reason for it:

    someresource { 'name':
      include => 'some value',<press C-R here>

It does fit the current regex because it both has include at the
beginning of the line only preceded by spaces and the line is ended with
a comma.

In order to avoid this extraneous indentation, let's just exclude cases
that contain either "=>" or "+>" on the line: we'll interpret this as
representing a parameter being set for a resource.