Preemptive return of indent for closing braces/parens (closes: #20 #24)
When we are closing a brace or a parenthenses on a line that follows a
comment line, the current code for GetPuppetIndent is returning an
indent too fast: it only checks that the previous line contains only a
comment, but does not acknowledge that the current line might need
further treatment.

In order to get the correct indentation in the case explained above we
need to skip the case that keeps the current indentation after a comment
line and let the function find out that we have a lone closing brace or
parenthenses.