searchpair improperly matching inside comments and strings
When trying to find out a matching brace or parentheses in order to
match its indent level, the current call to searchpair matches
characters inside comments and strings, which makes autoindentation
misinterpret where the matching brace or parentheses is positioned.

For example in the next case, searchpair would return -1 to indicate
failure to match and indent level would not get decreased:

    class something {
      #}
      }

In the above example, after typing the last curly brace, we expect
indentation to be brought back to the same column as the start of
"class", but it doesn't change.

The solution is to tell searchpair to ignore matches inside comments and
strings. To do this, we inspect the current syntax class that syntax
hilighting has determined at match point.