Avoid changing indent lines that are inside a multi-line string
When changing indentation inside of a multi-line string, we're actually
changing the contents of a string that will probably be output in a
config file on machines. This means that we're having an impact on
machines and not just on the code's presentation. This is bad and we
should avoid doing that.

Testing the syntax hilighting class that was found on first column of
the current line ensures that we're on a line that is continuing a
string: if a line starts with:

    "string"

then, the first column doesn't have the hilighting class yet. This means
that indentation of lines that contain only a string that starts on
column one will still be set properly.