Recover indent after multi-line string
Since the last commit is keeping text exactly in place, we can't relate
to pline anymore for aligning code properly anymore. For this we need to
find the last line that was not inside of a multi-line string and
reinitialize the indent level with the level that was set there.

If we don't do this it means code after a multi-line string could end up
aligning with any random indentation level.