PartOfInclude should not match resource parameters named "include"
This issue is similar to the one that was fixed in 5692a81: when trying
to identify the special case that looks like:

    include class1,
      class2

the function PartOfInclude matches too broadly and catches cases where a
resource parameter named "include" is present. This does not correspond
to the above and should thus not match.