unset options from ftplugin and indent when changing filetype
This should make this plugin play nicer with others by avoiding to leak
out options to other filetypes.

I've switched the first lines of ftplugin to the long option names in
order to make it clearer what they are, but most importantly that they
are being cleared out later by undo_ftplugin.

Finally, formatexpr gets moved over to ftplugin since it's not only
related to indent.