avoid breaking up long lines if textwidth is disabled
if users disable textwidth, they expect long lines not to get broken up
in smaller ones automatically.

also when textwidth is disabled, the test of whether the line is longer
than the value of &textwidth does always result in running the
formatting code in fallback for all non-empty lines.

the tests now belongs in a new category since we'll only see the effects
of formatting a long line now if we enable textwidth -- and it's not
linked to aligning hashrockets.