make syntax highlighting of epp simple but correct
This work is highly based off of Tim Pope's (and contributors') work on
the eruby syntax highlighting.

The eruby code automatically detects the underlying file type in order
to colorize the template content correctly for multiple types of
contents.
I didn't quite know how to implement this intelligently for .epp files
since most of the times we won't have a good clue what type of file the
contents represent.

As a first step, I've chosed to use "sh" for the underlying file type
since it's probably going to be the most common, and it is generic
enough that it won't look too bad for most config files.
We can eventually make this better. But simple steps first.