fix tests for new syntax highlighting
Some of the syntax group names have changed, and matching for numbers
has a better regexp that recognizes leading negative sign as being part
of the numbers.

There's one bug however: since curly braces are now recognized as the
end of a region, closing curly braces that are alone in a puppet code
section won't be matched properly. It's a bug that might need to be
addressed at some point, but for now let's disable that test.