Override prior type detection for embedded puppet (closes: #127)
Currently, if an `.epp` file starts with a code opening tag, `<%`, it gets
auto-detected as being of the `mason` type by Vim's core type detection,
which runs before ftdetect scripts are interpreted. The `.epp` extension
is not used by anything else, so we want to make sure that the core
autodetection doesn't take effect here.

To also document the other exception to file type detection in our
plugin, the use of `au!` is necessary for `.pp` file since the core type
detection already assigns them as pascal. Let's make a note as to why
we're using this forceful command in this particular case.

The effect of the `!` after the `au` keyword is that all autocommands get
removed and then this one gets added. This could potentially hide some
other autocommands that would be supplied by other plugins or user
scripts that get loaded before this plugin.

There should not be a need for an override of any type for Puppetfile,
so let's stop using the `!` suffix for that case.