Don't delete all autocmds when reading or opening a file
The effect of the ! after the au keyword is that all autocommands get
removed and then this one gets added.

This could potentially hide some other autocommands that would be
supplied by other plugins or user scripts that get loaded before this
plugin.

As far as I can see, other plugins that set the filetype don't use the !
after au.