fix CI builds for neovim
the travis-setup.sh is broken and hasn't been maintained for seven years
now.

I've looked up how coc.nvim is running their tests and I'm mimicking it
here so that we can have functional tests for nvim again.

coc.nvim uses a URL that says "nightly" instead of a version number.
That URL returns a 404. We'll have to stick with a versioned release for
nvim and bump the downloaded release once in a while.