check for code style with CI using vint
enforcing a sane code style is essential to keeping this codebase on a
higher level and avoid strange errors caused by anti-patterns.

the last "stable" release of vint happened in may 2019 and there's been
a couple of fixes applied since then, among which is to stop enforcing
strong comparison operators when the string contains nothing that can be
influenced by case matching. we'd like to use vint with those fixes