Use latest vim install action and run vader our own way
Apparently we're not using the latest version of the vim install action.
The version that we're using has a bug that was fixed in later releases.
As suggested in https://github.com/rhysd/action-setup-vim/issues/15,
using `@v1` instead of a specific version ensures that we're always
using the most up to date version in the 1.x.y space.

The vader action doesn't run properly and it does not have options for
specifying our own init script. So we're better off runing the tests
with the script that's already in the repository.